Add files via upload
First draft of problem 1, parts 1-3. Current problem is making the model for three. I'm trying to make a model where counts depend on the longitude of earth and have it look like it matches the scatter plot, but right now I think the exponential is too skewed. Possibly I shouldn't be using earth longitude with counts, or I solved for the model values incorrectly.